Rename DeltaUTxO modules to UTxOHistory.
